+++ SpookyJam2023/Assets/Scripts/Interfaces/IDamageable.cs
@@ -2,7 +2,7 @@ using UnityEngine.Events;
 
 public interface IDamageable
 {
-    public UnityAction<float> OnHealthChanged { get; set; }
+    public UnityAction<int> OnHealthChanged { get; set; }
     public UnityAction OnDead { get; set; }
 
     public void Damage(int damage);
